Fix issue submitting to Shaman from Windows machine
Fix a BAT error that arises when submitting a job from a Windows
workstation to a Shaman flamenco system. BAT expects an absolute path,
but on Windows systems, absolute paths begin with a drive letter.

This patch adds an arbitrary drive letter, and later removes it once the
relative paths have been computed.

Reviewed by: sybren

Differential Revision: https://developer.blender.org/D16084