Fix FFmpeg wanting to use JPEG files when rendering PNG
The preview video task would attempt to use JPEG preview files when the
"Preview" checkbox is checked, even when this checkbox is not shown in
Blender's UI (when the output format is not EXR). Blender still stores
this property, even when it's unused, and that's what tripped up the job
compiler.

The "Simple Blender Render" job type now first checks whether the previews
are necessary at all, and only then uses them.