Add Blender Conference 2023 to the website
For the hands-on class, there will be a special build of Flamenco. It
disables Shaman on all platforms (for uniformity for all participants)
and hard-codes `localhost` as URL for the Workers (to prevent
participants' Workers from finding each other's Managers).

The special build of Flamenco will be done on a separate branch
(`bcon-2023`), but the website should stay for longer, and thus this
commit is on the `main` branch.